requirements.txt
requests
pytest
jsonschema
allure-pytest
python-dotenv
python-dateutil
pytest-xdist